requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0
numpy>=1.24.0
matplotlib>=3.7.0
